Breaking change in PICA::Path
To further harmonize PICA Path expression with the syntax used in
pica-rs, the option position_as_occurrence is dropped, so a `/` always
starts occurrences instead of a position. Positions can still be used
after subfields, e.g. `123X$a/0-3` or `123X$*/0-3` (or `123X*/0-3` as
non-standard extension).

Reason: The slash is most used for occurrence so this syntax is easier
to use. Moreover a position makes more sense within a subfield than
across all subfields.

In addition a PICA Path expression without occurrence is equivalent to
one with occurrence zero, so `123X` will not match `123X/01`. Use `/*`
to match all occurrences.

Reason: The occurrence is part of the field identifier and missing
occurrence is internally equal to occurrence zero.